Remove content module from importers
