--- app/jobs/content/import_item_job.rb
@@ -1,9 +1,0 @@
-module Content
-  class ImportItemJob < ApplicationJob
-    sidekiq_options queue: 'publishing_api'
-
-    def run(*args)
-      Importers::SingleContentItem.run(*args)
-    end
-  end
-end
