+++ app/jobs/import_item_job.rb
@@ -1,0 +1,7 @@
+class ImportItemJob < ApplicationJob
+  sidekiq_options queue: 'publishing_api'
+
+  def run(*args)
+    Importers::SingleContentItem.run(*args)
+  end
+end
