+++ spec/domain/link_filter_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe Content::LinkFilter do
+RSpec.describe LinkFilter do
   subject { described_class }
   it "raises an error if constructing a filter with source and target" do
     expect {
@@ -16,14 +16,14 @@ RSpec.describe Content::LinkFilter do
     let!(:link) { create(:link, link_type: "organisations", source_content_id: "sid1", target_content_id: "tid1") }
 
     it "returns the links filter by source_ids" do
-      scope = Content::Item.all
+      scope = Item.all
       filter = subject.new(link_type: "organisations", source_ids: "sid1")
 
       expect(filter.apply(scope)).to match_array([content_item2])
     end
 
     it "returns the links filter by target ids" do
-      scope = Content::Item.all
+      scope = Item.all
       filter = subject.new(link_type: "organisations", target_ids: "tid1")
 
       expect(filter.apply(scope)).to match_array([content_item1])
